+++ src/services/client.js
@@ -14,10 +14,12 @@ const updateInfo = async (infoToUpdate, token) => {
       toast.success("Information updated!", {
         icon: "🚀",
       });
+      notify()
 
     return response.data;
   } catch (error) {
     const alertPassword = () => toast.error(error.message);
+    alertPassword()
   }
 };
